usercontroller tests completed

+++ src/main/java/com/peyman/blogapi/entity/repository/BlogRepository.java
@@ -5,7 +5,7 @@ import com.peyman.blogapi.entity.model.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface BlogRepository extends JpaRepository<Blog, Integer> {
+public interface BlogRepository extends JpaRepository<Blog, Long> {
 
 
     Blog getBlogByTitle(String title);
